# Store *[]byte in internal/repository.bufPool, not []byte