@fd0 2020-02-26
Interesting! I'm puzzled, what's going on here?

Where do the allocations come from, exactly?

If the pool is not needed (any more?), I'd rather remove it and make the code simpler and more robust instead of trying to squeeze every bit of speed out of it (after understanding what's going on here, of course). :)

It somehow feels wrong to me to save a pointer to the slice instead of the slice itself. We want to avoid allocating the (large) underlying array, so the slice structure itself is not of any interest.

If I understood [slices](https://blog.golang.org/go-slices-usage-and-internals) correctly, they consist of a pointer to the underlying array, a length and a cap field. Worst case if the structure for the slice needs to be reallocated each slice costs the runtime 24 byte (or so).