@fd0 2020-02-26
Ah! I found the [example](https://pkg.go.dev/sync?tab=doc#example-Pool) the comment in the code refers to, and I think I understood what's going on: if a non-pointer type is used, returning it from New() and Get() in the pool copies its value.