@MichaelEischer 2020-02-26
@fd0 `r.key.Seal` allocates a new array in the testcase as it has to handle a 4MB blob, whereas the buffer from the pool has only 8/3MB. That is the `ciphertext` slice returned by `r.key.Seal` uses a different backing buffer than the `ciphertext` slice that is passed in initially. `defer freeBuf(ciphertext)` captures a copy of the slice (just the descriptor, not the backing array) at the point in time the defer line is evaluated.

slices (that is the descriptor) in Go are by default passed by value. The only exception is when you explicitly get a pointer to a slice. In that case Go probably moves the slice from the stack to the heap. New() and Get() always return a copy of the value stored in the Pool. That can either be a copy of the slice (descriptor) or a copy of a pointer. By using pointers this PR changes the slices at which `*ciphertext` points. That way the call to defer will put the updated slice back into the pool.