@greatroar 2020-02-26
To elaborate on @MichaelEischer's point about descriptors, [slices](https://github.com/golang/go/blob/8bdd24df6afa746cede600351feb2aa4395a63ea/src/runtime/slice.go#L13) are syntactic sugar for

```
type slice struct {
        array unsafe.Pointer
        len   int
        cap   int
}
```

These are passed around by value.

An [interface](https://github.com/golang/go/blob/8bdd24df6afa746cede600351feb2aa4395a63ea/src/runtime/runtime2.go#L200) is actually a

```
type iface struct {
        tab  *itab
        data unsafe.Pointer
}
```

which can only hold a pointer. When you store a non-pointer value in an interface value, the address of the value is stored instead, potentially causing it to be moved to the heap (depending on escape analysis). So even when the pool prevents allocation of the slice's backing array, it still requires heap allocation of the slice header.

I'll remove the pool in a separate PR.